冰箱食物制作、物品标签

+++ src/main/java/com/xiaohunao/confluencedelight/common/init/ModItems.java
@@ -111,5 +111,5 @@ public class ModItems {
         return register(en, zh,1);
     }
 
-    public static Map<Item, Double> consumptionProbabilityMap;
+    public static Map<Item, Integer> consumptionProbabilityMap;
 }
